+++ atk-stm32f103/private_src/Serial.h
@@ -13,8 +13,8 @@ namespace bsp
 		int64_t Length() override;
 		void SetLength(int64_t value) override;
 
-		int64_t Read(uint8_t *buffer, int64_t offset, int64_t count) override;
-		void Write(uint8_t *buffer, int64_t offset, int64_t count) override;
+		int32_t Read(uint8_t *buffer, int32_t offset, int32_t count) override;
+		void Write(uint8_t *buffer, int32_t offset, int32_t count) override;
 
 		void Flush() override;
 		void Close() override;
@@ -24,7 +24,7 @@ namespace bsp
 
 		void Begin(uint32_t baud_rate) override;
 
-		int AvaliableToRead() override;
-		int AvaliableToWrite() override;
+		int32_t AvaliableToRead() override;
+		int32_t AvaliableToWrite() override;
 	};
 }
